Move paintLineOfSight() to ZeldaControls
