+++ app/main.js
@@ -86,7 +86,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_levels__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/levels */ \"./src/avo/levels.js\");\n/* harmony import */ var _avo_image_asset__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/image-asset */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_json_asset__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/json-asset */ \"./src/avo/json-asset.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\n\n\n\n\n\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar STARTING_LEVEL = Number.isInteger(parseInt(searchParams.get('level'))) ? parseInt(searchParams.get('level')) - 1 : 0;\n\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n\n    _classCallCheck(this, AvO);\n\n    var _args$width = args.width,\n        width = _args$width === void 0 ? 24 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n        _args$height = args.height,\n        height = _args$height === void 0 ? 16 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this.camera = {\n      target: null,\n      // Target atom to follow. If null, camera is static.\n      x: 0,\n      y: 0\n    };\n    this.setupUI();\n    this.initialised = false;\n    this.assets = {\n      \"exampleImage\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_3__[\"default\"]('assets/simple-bg.png'),\n      \"exampleJson\": new _avo_json_asset__WEBPACK_IMPORTED_MODULE_4__[\"default\"]('assets/example.json')\n    };\n    this.secretAssets = {// \"secretImage\": new ImageAsset('secrets/simple-bg.png'),\n      // \"secretJson\": new JsonAsset('secrets/example.json'),\n    };\n    this.hero = null;\n    this.atoms = [];\n    this.rules = {};\n    this.levels = new _avo_levels__WEBPACK_IMPORTED_MODULE_2__[\"default\"](this);\n    this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n    this.playerInput = {\n      // Mouse/touchscreen input\n      pointerStart: undefined,\n      pointerCurrent: undefined,\n      pointerEnd: undefined,\n      // Keys that are currently being pressed.\n      // keysPressed = { key: { duration, acknowledged } }\n      keysPressed: {}\n    };\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      }); // Paint status\n\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n\n      if (allAssetsReady) {\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        }); // Let's go!\n\n        this.initialised = true;\n        this.showUI();\n        this.levels.load(STARTING_LEVEL);\n      }\n    }\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n\n      if (this.initialised) {\n        this.play(timeStep);\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n    - timeStep: the time (milliseconds) since the last frame.\r\n      We expect 60 frames per second.\r\n     */\n\n  }, {\n    key: \"play\",\n    value: function play(timeStep) {\n      var _this2 = this;\n\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return; // Run the action gameplay\n      // ----------------\n\n      for (var id in this.rules) {\n        this.rules[id].play(timeStep);\n      }\n\n      this.atoms.forEach(function (atom) {\n        return atom.play(timeStep);\n      });\n      this.checkCollisions(timeStep); // Cleanup\n\n      this.atoms = this.atoms.filter(function (atom) {\n        return !atom._expired;\n      });\n\n      for (var _id in this.rules) {\n        if (this.rules[_id]._expired) delete this.rules[_id];\n      } // ----------------\n      // Increment the duration of each currently pressed key\n\n\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += timeStep;\n      });\n    }\n    /*\r\n    Draw a line of sight (cast a ray) starting from the hero, in the direction\r\n    of they are facing.\r\n     */\n\n  }, {\n    key: \"paintLineOfSight\",\n    value: function paintLineOfSight() {\n      if (!this.hero) return;\n      var hero = this.hero;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n      var MAX_LINE_OF_SIGHT_DISTANCE = 320; // Intended line of sight, i.e. a ray starting from the hero.\n\n      var lineOfSight = {\n        start: {\n          x: hero.x,\n          y: hero.y\n        },\n        end: {\n          x: hero.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(hero.rotation),\n          y: hero.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(hero.rotation)\n        }\n      };\n      var actualLineOfSightEndPoint = undefined; // For each atom, see if it intersects with the hero's LOS\n\n      this.atoms.forEach(function (atom) {\n        if (atom === hero) return; // TODO: check for opaqueness and/or if the atom is visible.\n\n        var vertices = atom.vertices;\n        if (vertices.length < 2) return; // Every atom has a \"shape\" that can be represented by a polygon.\n        // (Yes, even circles.) Check each segment (aka edge aka side) of the\n        // polygon.\n\n        for (var i = 0; i < vertices.length; i++) {\n          var segment = {\n            start: {\n              x: vertices[i].x,\n              y: vertices[i].y\n            },\n            end: {\n              x: vertices[(i + 1) % vertices.length].x,\n              y: vertices[(i + 1) % vertices.length].y\n            }\n          }; // Find the intersection. We want to find the intersection point\n          // closest to the hero (the LOS ray's starting point).\n\n          var intersection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].getLineIntersection(lineOfSight, segment);\n\n          if (!actualLineOfSightEndPoint || intersection && intersection.distanceFactor < actualLineOfSightEndPoint.distanceFactor) {\n            actualLineOfSightEndPoint = intersection;\n          }\n        }\n      });\n\n      if (!actualLineOfSightEndPoint) {\n        actualLineOfSightEndPoint = {\n          x: hero.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(hero.rotation),\n          y: hero.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(hero.rotation)\n        };\n      } // Expected line of sight\n\n\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x + camera.x, lineOfSight.start.y + camera.y);\n      c2d.lineTo(lineOfSight.end.x + camera.x, lineOfSight.end.y + camera.y);\n      c2d.closePath();\n      c2d.strokeStyle = '#c88';\n      c2d.lineWidth = 3;\n      c2d.setLineDash([5, 5]);\n      c2d.stroke();\n      c2d.setLineDash([]); // Actual line of sight\n\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x + camera.x, lineOfSight.start.y + camera.y);\n      c2d.lineTo(actualLineOfSightEndPoint.x + camera.x, actualLineOfSightEndPoint.y + camera.y);\n      c2d.closePath();\n      c2d.strokeStyle = '#39f';\n      c2d.lineWidth = 3;\n      c2d.stroke(); // Expected end of line of sight\n\n      c2d.beginPath();\n      c2d.arc(lineOfSight.end.x + camera.x, lineOfSight.end.y + camera.y, 4, 0, 2 * Math.PI);\n      c2d.closePath();\n      c2d.fillStyle = '#c88';\n      c2d.fill(); // Actual end of line of sight\n\n      c2d.beginPath();\n      c2d.arc(actualLineOfSightEndPoint.x + camera.x, actualLineOfSightEndPoint.y + camera.y, 8, 0, 2 * Math.PI);\n      c2d.closePath();\n      c2d.fillStyle = '#39f';\n      c2d.fill();\n    }\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n\n      var c2d = this.canvas2d;\n      var camera = this.camera; // Camera Controls: focus the camera on the target atom, if any.\n      // ----------------\n\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x;\n        camera.y = this.canvasHeight / 2 - camera.target.y;\n      }\n\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2; // ----------------\n      // Draw grid\n      // ----------------\n\n      var offsetX = this.camera.x % _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE - _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      var offsetY = this.camera.y % _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE - _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n\n      for (var y = offsetY; y < this.canvasHeight; y += _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n          c2d.stroke(); // Debug Grid\n\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"1em Source Code Pro\";\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n            var row = Math.floor((y - this.camera.y) / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n            c2d.fillText(col + ',' + row, x + _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 2, y + _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      } // ----------------\n      // Draw atoms and other elements\n      // ----------------\n\n\n      var _loop = function _loop(layer) {\n        _this3.atoms.forEach(function (atom) {\n          return atom.paint(layer);\n        });\n\n        for (var id in _this3.rules) {\n          _this3.rules[id].paint(layer);\n        }\n      };\n\n      for (var layer = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      } // ----------------\n      // Draw player input\n      // ----------------\n\n\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN && this.hero && this.playerInput.pointerCurrent) {\n        var inputCoords = this.playerInput.pointerCurrent;\n        c2d.strokeStyle = '#888';\n        c2d.lineWidth = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 8;\n        c2d.beginPath();\n        c2d.arc(inputCoords.x, inputCoords.y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, 0, 2 * Math.PI);\n        c2d.stroke();\n      } // ----------------\n\n\n      this.paintLineOfSight();\n    }\n    /*\r\n    Section: UI and Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      } // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n\n\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n\n\n        interactionMenu.load(div); // load the new menu\n\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.html.main.focus();\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      }\n\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n        // DEBUG\n\n        case 'z':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_5__[\"default\"](this));\n          }\n\n          break;\n      } // General input\n\n\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      this.playerInput.keysPressed[e.key] = undefined;\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var isFullscreen = document.fullscreenElement;\n\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.className = 'fullscreen';\n          this.html.main.requestFullscreen();\n        }\n      } else {\n        var _document$exitFullscr, _document;\n\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 ? void 0 : _document$exitFullscr.call(_document);\n        this.html.main.className = '';\n      }\n\n      this.updateUI();\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      this.levels.reload();\n    }\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules[id] = rule;\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      for (var id in this.rules) {\n        delete this.rules[id];\n      }\n    }\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions(timeStep) {\n      for (var a = 0; a < this.atoms.length; a++) {\n        var atomA = this.atoms[a];\n\n        for (var b = a + 1; b < this.atoms.length; b++) {\n          var atomB = this.atoms[b];\n          var collisionCorrection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(atomA, atomB);\n\n          if (collisionCorrection) {\n            atomA.onCollision(atomB, collisionCorrection.a);\n            atomB.onCollision(atomA, collisionCorrection.b);\n          }\n        }\n      }\n    }\n  }]);\n\n  return AvO;\n}();\n\n\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\n\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_levels__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/levels */ \"./src/avo/levels.js\");\n/* harmony import */ var _avo_image_asset__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/image-asset */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_json_asset__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/json-asset */ \"./src/avo/json-asset.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\n\n\n\n\n\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar STARTING_LEVEL = Number.isInteger(parseInt(searchParams.get('level'))) ? parseInt(searchParams.get('level')) - 1 : 0;\n\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n\n    _classCallCheck(this, AvO);\n\n    var _args$width = args.width,\n        width = _args$width === void 0 ? 24 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n        _args$height = args.height,\n        height = _args$height === void 0 ? 16 * _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this.camera = {\n      target: null,\n      // Target atom to follow. If null, camera is static.\n      x: 0,\n      y: 0\n    };\n    this.setupUI();\n    this.initialised = false;\n    this.assets = {\n      \"exampleImage\": new _avo_image_asset__WEBPACK_IMPORTED_MODULE_3__[\"default\"]('assets/simple-bg.png'),\n      \"exampleJson\": new _avo_json_asset__WEBPACK_IMPORTED_MODULE_4__[\"default\"]('assets/example.json')\n    };\n    this.secretAssets = {// \"secretImage\": new ImageAsset('secrets/simple-bg.png'),\n      // \"secretJson\": new JsonAsset('secrets/example.json'),\n    };\n    this.hero = null;\n    this.atoms = [];\n    this.rules = {};\n    this.levels = new _avo_levels__WEBPACK_IMPORTED_MODULE_2__[\"default\"](this);\n    this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n    this.playerInput = {\n      // Mouse/touchscreen input\n      pointerStart: undefined,\n      pointerCurrent: undefined,\n      pointerEnd: undefined,\n      // Keys that are currently being pressed.\n      // keysPressed = { key: { duration, acknowledged } }\n      keysPressed: {}\n    };\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      }); // Paint status\n\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n\n      if (allAssetsReady) {\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        }); // Let's go!\n\n        this.initialised = true;\n        this.showUI();\n        this.levels.load(STARTING_LEVEL);\n      }\n    }\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n\n      if (this.initialised) {\n        this.play(timeStep);\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n    - timeStep: the time (milliseconds) since the last frame.\r\n      We expect 60 frames per second.\r\n     */\n\n  }, {\n    key: \"play\",\n    value: function play(timeStep) {\n      var _this2 = this;\n\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return; // Run the action gameplay\n      // ----------------\n\n      for (var id in this.rules) {\n        this.rules[id].play(timeStep);\n      }\n\n      this.atoms.forEach(function (atom) {\n        return atom.play(timeStep);\n      });\n      this.checkCollisions(timeStep); // Cleanup\n\n      this.atoms = this.atoms.filter(function (atom) {\n        return !atom._expired;\n      });\n\n      for (var _id in this.rules) {\n        if (this.rules[_id]._expired) delete this.rules[_id];\n      } // ----------------\n      // Increment the duration of each currently pressed key\n\n\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += timeStep;\n      });\n    }\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n\n      var c2d = this.canvas2d;\n      var camera = this.camera; // Camera Controls: focus the camera on the target atom, if any.\n      // ----------------\n\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x;\n        camera.y = this.canvasHeight / 2 - camera.target.y;\n      }\n\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2; // ----------------\n      // Draw grid\n      // ----------------\n\n      var offsetX = this.camera.x % _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE - _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n      var offsetY = this.camera.y % _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE - _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE;\n\n      for (var y = offsetY; y < this.canvasHeight; y += _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n          c2d.stroke(); // Debug Grid\n\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"1em Source Code Pro\";\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n            var row = Math.floor((y - this.camera.y) / _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n            c2d.fillText(col + ',' + row, x + _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 2, y + _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      } // ----------------\n      // Draw atoms and other elements\n      // ----------------\n\n\n      var _loop = function _loop(layer) {\n        _this3.atoms.forEach(function (atom) {\n          return atom.paint(layer);\n        });\n\n        for (var id in _this3.rules) {\n          _this3.rules[id].paint(layer);\n        }\n      };\n\n      for (var layer = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      } // ----------------\n      // Draw player input\n      // ----------------\n\n\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN && this.hero && this.playerInput.pointerCurrent) {\n        var inputCoords = this.playerInput.pointerCurrent;\n        c2d.strokeStyle = '#888';\n        c2d.lineWidth = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE / 8;\n        c2d.beginPath();\n        c2d.arc(inputCoords.x, inputCoords.y, _avo_constants__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, 0, 2 * Math.PI);\n        c2d.stroke();\n      } // ----------------\n\n    }\n    /*\r\n    Section: UI and Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      } // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n\n\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n\n\n        interactionMenu.load(div); // load the new menu\n\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.html.main.focus();\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n\n      if (this.playerAction === _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerAction = _avo_constants__WEBPACK_IMPORTED_MODULE_0__.PLAYER_ACTIONS.IDLE;\n      }\n\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n        // DEBUG\n\n        case 'z':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_5__[\"default\"](this));\n          }\n\n          break;\n      } // General input\n\n\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      this.playerInput.keysPressed[e.key] = undefined;\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var isFullscreen = document.fullscreenElement;\n\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.className = 'fullscreen';\n          this.html.main.requestFullscreen();\n        }\n      } else {\n        var _document$exitFullscr, _document;\n\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 ? void 0 : _document$exitFullscr.call(_document);\n        this.html.main.className = '';\n      }\n\n      this.updateUI();\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      this.levels.reload();\n    }\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules[id] = rule;\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      for (var id in this.rules) {\n        delete this.rules[id];\n      }\n    }\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions(timeStep) {\n      for (var a = 0; a < this.atoms.length; a++) {\n        var atomA = this.atoms[a];\n\n        for (var b = a + 1; b < this.atoms.length; b++) {\n          var atomB = this.atoms[b];\n          var collisionCorrection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(atomA, atomB);\n\n          if (collisionCorrection) {\n            atomA.onCollision(atomB, collisionCorrection.a);\n            atomB.onCollision(atomA, collisionCorrection.b);\n          }\n        }\n      }\n    }\n  }]);\n\n  return AvO;\n}();\n\n\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\n\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
 
 /***/ }),
 
@@ -216,7 +216,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \**********************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ZeldaControls)\n/* harmony export */ });\n/* harmony import */ var _avo_rule__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/rule */ \"./src/avo/rule/index.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n/*\r\nStandard player controls for top-down adventure games.\r\n */\n\nvar ZeldaControls = /*#__PURE__*/function (_Rule) {\n  _inherits(ZeldaControls, _Rule);\n\n  var _super = _createSuper(ZeldaControls);\n\n  function ZeldaControls(app) {\n    var _this;\n\n    _classCallCheck(this, ZeldaControls);\n\n    _this = _super.call(this, app);\n    _this._type = 'zelda-controls';\n    return _this;\n  }\n\n  _createClass(ZeldaControls, [{\n    key: \"play\",\n    value: function play(timeStep) {\n      var app = this._app;\n\n      _get(_getPrototypeOf(ZeldaControls.prototype), \"play\", this).call(this, timeStep);\n\n      if (app.hero) {\n        var keysPressed = app.playerInput.keysPressed;\n        var intent = undefined;\n        var directionX = 0;\n        var directionY = 0;\n        if (keysPressed['ArrowRight']) directionX++;\n        if (keysPressed['ArrowDown']) directionY++;\n        if (keysPressed['ArrowLeft']) directionX--;\n        if (keysPressed['ArrowUp']) directionY--;\n\n        if (keysPressed['x'] && !keysPressed['x'].acknowledged || keysPressed['X'] && !keysPressed['X'].acknowledged) {\n          intent = {\n            name: 'dash',\n            directionX: directionX,\n            directionY: directionY\n          };\n          if (keysPressed['x']) keysPressed['x'].acknowledged = true;\n          if (keysPressed['X']) keysPressed['X'].acknowledged = true;\n        } else if (directionX || directionY) {\n          intent = {\n            name: 'move',\n            directionX: directionX,\n            directionY: directionY\n          };\n        }\n\n        app.hero.intent = intent;\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var hero = this._app.hero;\n      var c2d = this._app.canvas2d;\n      var camera = this._app.camera;\n\n      if (layer === _avo_constants__WEBPACK_IMPORTED_MODULE_1__.LAYERS.HUD) {\n        var _hero$action;\n\n        // Draw UI data\n        // ----------------\n        var X_OFFSET = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE * 1.5;\n        var Y_OFFSET = _avo_constants__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE * -1.0;\n        var LEFT = X_OFFSET;\n        var RIGHT = this._app.canvasWidth - X_OFFSET;\n        var BOTTOM = this._app.canvasHeight + Y_OFFSET;\n        c2d.font = '3em Source Code Pro';\n        c2d.textBaseline = 'bottom';\n        c2d.lineWidth = 8;\n        var health = Math.max(hero === null || hero === void 0 ? void 0 : hero.health, 0) || 0;\n        var text = '❤️'.repeat(health);\n        c2d.textAlign = 'left';\n        c2d.strokeStyle = '#fff';\n        c2d.strokeText(text, LEFT, BOTTOM);\n        c2d.fillStyle = '#c44';\n        c2d.fillText(text, LEFT, BOTTOM);\n        text = (hero === null || hero === void 0 ? void 0 : (_hero$action = hero.action) === null || _hero$action === void 0 ? void 0 : _hero$action.name) + ' (' + (hero === null || hero === void 0 ? void 0 : hero.moveSpeed.toFixed(2)) + ')';\n        c2d.textAlign = 'right';\n        c2d.strokeStyle = '#fff';\n        c2d.strokeText(text, RIGHT, BOTTOM);\n        c2d.fillStyle = '#c44';\n        c2d.fillText(text, RIGHT, BOTTOM); // ----------------\n      }\n    }\n  }]);\n\n  return ZeldaControls;\n}(_avo_rule__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/rule/types/zelda-controls.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ZeldaControls)\n/* harmony export */ });\n/* harmony import */ var _avo_rule__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/rule */ \"./src/avo/rule/index.js\");\n/* harmony import */ var _avo_physics__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_constants__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/constants */ \"./src/avo/constants.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\n\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n/*\r\nStandard player controls for top-down adventure games.\r\n */\n\nvar ZeldaControls = /*#__PURE__*/function (_Rule) {\n  _inherits(ZeldaControls, _Rule);\n\n  var _super = _createSuper(ZeldaControls);\n\n  function ZeldaControls(app) {\n    var _this;\n\n    _classCallCheck(this, ZeldaControls);\n\n    _this = _super.call(this, app);\n    _this._type = 'zelda-controls';\n    return _this;\n  }\n\n  _createClass(ZeldaControls, [{\n    key: \"play\",\n    value: function play(timeStep) {\n      var app = this._app;\n      var hero = app.hero;\n\n      _get(_getPrototypeOf(ZeldaControls.prototype), \"play\", this).call(this, timeStep);\n\n      if (hero) {\n        var keysPressed = app.playerInput.keysPressed;\n        var intent = undefined;\n        var directionX = 0;\n        var directionY = 0;\n        if (keysPressed['ArrowRight']) directionX++;\n        if (keysPressed['ArrowDown']) directionY++;\n        if (keysPressed['ArrowLeft']) directionX--;\n        if (keysPressed['ArrowUp']) directionY--;\n\n        if (keysPressed['x'] && !keysPressed['x'].acknowledged || keysPressed['X'] && !keysPressed['X'].acknowledged) {\n          intent = {\n            name: 'dash',\n            directionX: directionX,\n            directionY: directionY\n          };\n          if (keysPressed['x']) keysPressed['x'].acknowledged = true;\n          if (keysPressed['X']) keysPressed['X'].acknowledged = true;\n        } else if (directionX || directionY) {\n          intent = {\n            name: 'move',\n            directionX: directionX,\n            directionY: directionY\n          };\n        }\n\n        hero.intent = intent;\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var hero = this._app.hero;\n      var c2d = this._app.canvas2d;\n      var camera = this._app.camera;\n\n      if (layer === _avo_constants__WEBPACK_IMPORTED_MODULE_2__.LAYERS.HUD) {\n        var _hero$action;\n\n        // Draw UI data\n        // ----------------\n        var X_OFFSET = _avo_constants__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 1.5;\n        var Y_OFFSET = _avo_constants__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * -1.0;\n        var LEFT = X_OFFSET;\n        var RIGHT = this._app.canvasWidth - X_OFFSET;\n        var BOTTOM = this._app.canvasHeight + Y_OFFSET;\n        c2d.font = '3em Source Code Pro';\n        c2d.textBaseline = 'bottom';\n        c2d.lineWidth = 8;\n        var health = Math.max(hero === null || hero === void 0 ? void 0 : hero.health, 0) || 0;\n        var text = '❤️'.repeat(health);\n        c2d.textAlign = 'left';\n        c2d.strokeStyle = '#fff';\n        c2d.strokeText(text, LEFT, BOTTOM);\n        c2d.fillStyle = '#c44';\n        c2d.fillText(text, LEFT, BOTTOM);\n        text = (hero === null || hero === void 0 ? void 0 : (_hero$action = hero.action) === null || _hero$action === void 0 ? void 0 : _hero$action.name) + ' (' + (hero === null || hero === void 0 ? void 0 : hero.moveSpeed.toFixed(2)) + ')';\n        c2d.textAlign = 'right';\n        c2d.strokeStyle = '#fff';\n        c2d.strokeText(text, RIGHT, BOTTOM);\n        c2d.fillStyle = '#c44';\n        c2d.fillText(text, RIGHT, BOTTOM); // ----------------\n\n        this.paintLineOfSight(hero);\n      }\n    }\n    /*\r\n    Draw a line of sight (cast a ray) starting from a specified Atom (usually the\r\n    hero), in the direction they're facing.\r\n     */\n\n  }, {\n    key: \"paintLineOfSight\",\n    value: function paintLineOfSight(srcAtom) {\n      if (!srcAtom) return;\n      var c2d = this._app.canvas2d;\n      var camera = this._app.camera;\n      var atoms = this._app.atoms;\n      var MAX_LINE_OF_SIGHT_DISTANCE = 320; // Intended line of sight, i.e. a ray starting from the hero/source Atom.\n\n      var lineOfSight = {\n        start: {\n          x: srcAtom.x,\n          y: srcAtom.y\n        },\n        end: {\n          x: srcAtom.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcAtom.rotation),\n          y: srcAtom.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcAtom.rotation)\n        }\n      };\n      var actualLineOfSightEndPoint = undefined; // For each other Atom, see if it intersects with the source Atom's LOS\n\n      atoms.forEach(function (atom) {\n        if (atom === srcAtom) return; // TODO: check for opaqueness and/or if the atom is visible.\n\n        var vertices = atom.vertices;\n        if (vertices.length < 2) return; // Every atom has a \"shape\" that can be represented by a polygon.\n        // (Yes, even circles.) Check each segment (aka edge aka side) of the\n        // polygon.\n\n        for (var i = 0; i < vertices.length; i++) {\n          var segment = {\n            start: {\n              x: vertices[i].x,\n              y: vertices[i].y\n            },\n            end: {\n              x: vertices[(i + 1) % vertices.length].x,\n              y: vertices[(i + 1) % vertices.length].y\n            }\n          }; // Find the intersection. We want to find the intersection point\n          // closest to the source Atom (the LOS ray's starting point).\n\n          var intersection = _avo_physics__WEBPACK_IMPORTED_MODULE_1__[\"default\"].getLineIntersection(lineOfSight, segment);\n\n          if (!actualLineOfSightEndPoint || intersection && intersection.distanceFactor < actualLineOfSightEndPoint.distanceFactor) {\n            actualLineOfSightEndPoint = intersection;\n          }\n        }\n      });\n\n      if (!actualLineOfSightEndPoint) {\n        actualLineOfSightEndPoint = {\n          x: srcAtom.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcAtom.rotation),\n          y: srcAtom.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcAtom.rotation)\n        };\n      } // Expected line of sight\n\n\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x + camera.x, lineOfSight.start.y + camera.y);\n      c2d.lineTo(lineOfSight.end.x + camera.x, lineOfSight.end.y + camera.y);\n      c2d.closePath();\n      c2d.strokeStyle = '#c88';\n      c2d.lineWidth = 3;\n      c2d.setLineDash([5, 5]);\n      c2d.stroke();\n      c2d.setLineDash([]); // Actual line of sight\n\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x + camera.x, lineOfSight.start.y + camera.y);\n      c2d.lineTo(actualLineOfSightEndPoint.x + camera.x, actualLineOfSightEndPoint.y + camera.y);\n      c2d.closePath();\n      c2d.strokeStyle = '#39f';\n      c2d.lineWidth = 3;\n      c2d.stroke(); // Expected end of line of sight\n\n      c2d.beginPath();\n      c2d.arc(lineOfSight.end.x + camera.x, lineOfSight.end.y + camera.y, 4, 0, 2 * Math.PI);\n      c2d.closePath();\n      c2d.fillStyle = '#c88';\n      c2d.fill(); // Actual end of line of sight\n\n      c2d.beginPath();\n      c2d.arc(actualLineOfSightEndPoint.x + camera.x, actualLineOfSightEndPoint.y + camera.y, 8, 0, 2 * Math.PI);\n      c2d.closePath();\n      c2d.fillStyle = '#39f';\n      c2d.fill();\n    }\n  }]);\n\n  return ZeldaControls;\n}(_avo_rule__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/rule/types/zelda-controls.js?");
 
 /***/ }),
